Remove comments tree view object
