+++ app/helpers/comments_helper.rb
@@ -12,6 +12,16 @@ module CommentsHelper
   end
 
   def tree_for(comment, sub_comments, commentable)
-    Comments::Tree.new(context: self, comment: comment, sub_comments: sub_comments, commentable: commentable).display
+    nested_comments(tree: { comment => sub_comments }, commentable: commentable, is_view_root: true)
+  end
+
+  private
+
+  def nested_comments(tree:, commentable:, is_view_root: false)
+    tree.map do |comment, sub_comments|
+      render("comments/comment", comment: comment, commentable: commentable,
+                                 is_view_root: is_view_root, is_childless: sub_comments.empty?,
+                                 subtree_html: nested_comments(tree: sub_comments, commentable: commentable))
+    end.join.html_safe
   end
 end
